yarsu-telegram-bot/pyproject.toml
[project]
name = "yarsu-telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "python-telegram-bot>=22.5",
]
